refactored and added storybook
