+++ App.tsx
@@ -1,4 +1,4 @@
-import { StyleSheet, Text, View } from 'react-native';
+import { StyleSheet, View } from 'react-native';
 import Location from './app/Location'
 
 export default function App() {
