+++ README.md
@@ -1,4 +1,4 @@
-### bag-reminder
+# Bag Reminder
 
 ## Latest
 ![qr](https://github.com/rsawyerdev/bag-reminder/assets/1273821/0dd7337f-b717-4b58-98bc-8833d2419ed4)
